fix: persist Yellow Card deposits/withdrawals and reconcile via webhook
Deposits and withdrawals were only ever tracked as a Redis idempotency
cache entry and the initial API response — nothing was written to
Postgres, and there was no way for Yellow Card to notify us of a status
change after the fact. A crash, restart, or delayed bank transfer could
leave the platform's view of a transaction permanently diverged from
Yellow Card's.

- Add a `deposit` domain package (model + repository) and finish wiring
  up the previously dead-code `withdrawal` package (its repository
  referenced a `withdrawals` table that didn't exist in any migration).
- Migration 038 creates `deposits` and `withdrawals` tables, unique on
  payment_ref/receive_id/yellow_card_tx_id for webhook lookups.
- DepositHandler now persists a row right after Yellow Card accepts a
  receive/send request; a persistence failure is a hard error back to
  the caller (with the receive/send ID logged) rather than a silently
  untracked transaction.
- New YellowCardWebhookHandler (POST /webhooks/yellowcard) verifies the
  HMAC signature and reconciles the matching deposit/withdrawal against
  the reported status (completed/failed/processing).
- Switched withdrawal amounts from int64 to float64 to match the
  float-based amounts used throughout the Yellow Card client/handlers.
- Wired WithRedis/WithConfig into the production DepositHandler — they
  were defined but never called, so daily caps and idempotency caching
  were silently inactive.
- Fixed a test bug (context key "user_id" vs "userID") that had been
  masking a broken daily-cap-limit test.

Verified migration 038 up/down and the new repositories end-to-end
against a local Postgres instance.

+++ internal/domain/withdrawal/model.go
@@ -14,37 +14,37 @@ const (
 )
 
 type Withdrawal struct {
-	ID              string          `json:"id" db:"id"`
-	UserID          string          `json:"userId" db:"user_id"`
-	AmountUSDC      int64           `json:"amountUsdc" db:"amount_usdc"`
-	EstimatedNGN    int64           `json:"estimatedNgn" db:"estimated_ngn"`
-	BankCode        string          `json:"bankCode" db:"bank_code"`
-	AccountNumber   string          `json:"accountNumber" db:"account_number"`
-	AccountName     string          `json:"accountName" db:"account_name"`
+	ID              string           `json:"id" db:"id"`
+	UserID          string           `json:"userId" db:"user_id"`
+	AmountUSDC      float64          `json:"amountUsdc" db:"amount_usdc"`
+	EstimatedNGN    float64          `json:"estimatedNgn" db:"estimated_ngn"`
+	BankCode        string           `json:"bankCode" db:"bank_code"`
+	AccountNumber   string           `json:"accountNumber" db:"account_number"`
+	AccountName     string           `json:"accountName" db:"account_name"`
 	Status          WithdrawalStatus `json:"status" db:"status"`
-	
+
 	// Platform receiving address (where user sends USDC)
-	PlatformAddress string          `json:"platformAddress" db:"platform_address"`
-	
+	PlatformAddress string `json:"platformAddress" db:"platform_address"`
+
 	// Transaction hashes
-	USDCtxHash      *string         `json:"usdcTxHash,omitempty" db:"usdc_tx_hash"`
-	YellowCardTxID  *string         `json:"yellowCardTxId,omitempty" db:"yellow_card_tx_id"`
-	
+	USDCtxHash     *string `json:"usdcTxHash,omitempty" db:"usdc_tx_hash"`
+	YellowCardTxID *string `json:"yellowCardTxId,omitempty" db:"yellow_card_tx_id"`
+
 	// Timestamps
-	CreatedAt       time.Time       `json:"createdAt" db:"created_at"`
-	ReceivedAt      *time.Time      `json:"receivedAt,omitempty" db:"received_at"`
-	CompletedAt     *time.Time      `json:"completedAt,omitempty" db:"completed_at"`
-	
+	CreatedAt   time.Time  `json:"createdAt" db:"created_at"`
+	ReceivedAt  *time.Time `json:"receivedAt,omitempty" db:"received_at"`
+	CompletedAt *time.Time `json:"completedAt,omitempty" db:"completed_at"`
+
 	// Error handling
-	FailureReason   *string         `json:"failureReason,omitempty" db:"failure_reason"`
-	
+	FailureReason *string `json:"failureReason,omitempty" db:"failure_reason"`
+
 	// Yellow Card details
-	PaymentRef      string          `json:"paymentRef" db:"payment_ref"`
+	PaymentRef string `json:"paymentRef" db:"payment_ref"`
 }
 
 type WithdrawalRequest struct {
-	AmountUSDC    int64  `json:"amountUsdc" binding:"required,gt=0"`
-	BankCode      string `json:"bankCode" binding:"required"`
-	AccountNumber string `json:"accountNumber" binding:"required"`
-	AccountName   string `json:"accountName" binding:"required"`
+	AmountUSDC    float64 `json:"amountUsdc" binding:"required,gt=0"`
+	BankCode      string  `json:"bankCode" binding:"required"`
+	AccountNumber string  `json:"accountNumber" binding:"required"`
+	AccountName   string  `json:"accountName" binding:"required"`
 }
